Time: 16 ms (16.23%) | Memory: 8.1 MB (96.32%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -3,7 +3,7 @@ public:
     bool isPalindrome(int x) {
         
         long revNum = 0;
-        long dup = x;
+        int dup = x;
         
         while(x>0){
             int lastDigit = x%10;
